ExtractAggregatedHealth and add test
